refactor: membercard update & change file structure
