+++ styles/MemberCard.module.css
@@ -28,10 +28,6 @@
   gap: 8px;
 }
 
-.card__title_container_column_reverse h1 {
-  font-size: 32px;
-}
-
 .card__image_large {
   width: 100%;
 }
